drm/i915/gsc: define gsc fw
Add FW definition and the matching override modparam.

The GSC FW has both a release version, based on platform and a rolling
counter, and a compatibility version, which is the one tracking
interface changes. Since what we care about is the interface, we use
the compatibility version in the binary names.

Same as with the GuC, a major version bump indicate a
backward-incompatible change, while a minor version bump indicates a
backward-compatible one, so we use only the former in the file name.

topic/core-for-CI note: the GSC team has asked us to hold off sending
the GSC FW to the linux-firmware repo, so we can't merge this to
drm-intel-gt-next yet. However, we do want to start running the GSC
code in CI, hence merging to topic/core-for-CI for now. See the gitlab
issue for expected FW availability.

[merge note: drm-intel-gt-next has the FW defs outside of the autoselect
function, so this patch has been updated to match to reduce conflicts
when building tip]

References: https://gitlab.freedesktop.org/drm/intel/-/issues/8705
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Alan Previn <alan.previn.teres.alexis@intel.com>
Cc: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Acked-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230628173708.3239698-1-daniele.ceraolospurio@intel.com